TRANSIMS Version 6.0
Router 6.0.40
Logic was added to identify and correct negative start times before the time sorting index is created to avoid sort problems in the output plan file.

Simulator 6.0.4
Logic was added to properly record entries from and exits to parking lots in the performance file entry and exit fields.  Corrections were made to the way capacity factors are implemented in macroscopic simulations.