TRANSIMS Version 6.0 – December 2014
SysLib
Additional options were added to the Performance processing codes.

File_Data_Service
The vehicle hours of delay calculation within the Performance data structure was modified to apply flow rather than volume to the link travel time difference.  This makes the delay values consistent with the vehicle hours of travel calculations.

ArcPerf 6.0.11
The vehicle hours of delay calculations were updated.

LinkSum 6.0.27
The vehicle hours of day calculations were updated.  Logic was added to zero out comparison cells for links/directions that are not included in both networks.  The link mapping logic for performance comparisons was expanded to permit multiple links to be mapped to a single links.

NetMerge 6.0.0
New program to merge two TRANSIMS networks together.

PathSkim 6.0.10
NEW_ACCESSIBILITY_FILE/FORMAT, ORIGIN_WEIGHT_FIELD, DESTINATION_WEIGHT_FIELD, and MAXIMUM_TRAVEL_TIME keys were added to generate zone-based summarizes of travel time accessibility measures based on origin and destination size values.  This is typically used to calculate the number of jobs that can be reached from households in a given zone within a specified travel times by auto or transit.

PerfPrep 6.0.10
The processing options and the travel time ratio checks were expanded.  The vehicle hours of delay calculation was updated.

SubareaNet 6.0.8
The SAVE_OUTSIDE_NETWORK key was added.  This option extracts a network outside of the subarea boundary that can be merged with subarea data to reconstruct a regional network with network changes or results within the subarea.

SubareaPlans 6.0.2
Logic was added to detect situations where links in the plan file are not included in the subarea network file.