TRANSIMS Version 6.0
Execution_Service
A bug was fixed in the report appending logic.

Simulator_Service
Additional flexibility was added to compile with the ROUTING option, but to execute the program without routing.  Additional CHECK messages were added to detect problems with array index problems.

Simulator 6.0.18
Plan and Trip file processing was aligned with the ROUTING compile key and control key options to enable a program compiled with ROUTING activated to enable the option to skip the routing functionality and read an input plan file.